+++ src/fife/eventchannel/key/ikeylistener.h
@@ -8,10 +8,8 @@
 #include "platform.h"
 
 // Standard C++ library includes
-//
 
 // 3rd party library includes
-//
 
 // FIFE includes
 #include "eventchannel/base/ilistener.h"